+++ src/DotCommon/DotCommon/IO/FileHelper.cs
@@ -1,18 +1,27 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 
 namespace DotCommon.IO
 {
     /// <summary>
-    /// 文件工具类
+    /// Provides utility methods for file operations.
     /// </summary>
     public static class FileHelper
     {
         /// <summary>
-        /// 如果文件存在就删除文件
+        /// Deletes a file if it exists.
         /// </summary>
-        /// <param name="fileName">文件名</param>
+        /// <param name="fileName">The name of the file to delete.</param>
+        /// <exception cref="ArgumentNullException">Thrown when fileName is null.</exception>
+        /// <exception cref="ArgumentException">Thrown when fileName is empty or contains invalid characters.</exception>
         public static void DeleteIfExists(string fileName)
         {
+            if (fileName == null)
+                throw new ArgumentNullException(nameof(fileName));
+
+            if (string.IsNullOrEmpty(fileName))
+                throw new ArgumentException("File name cannot be null or empty.", nameof(fileName));
+
             if (File.Exists(fileName))
             {
                 File.Delete(fileName);
@@ -20,12 +29,21 @@ namespace DotCommon.IO
         }
 
         /// <summary>
-        /// 获取文件的大小
+        /// Gets the size of a file.
         /// </summary>
-        /// <param name="fileName"></param>
-        /// <returns></returns>
+        /// <param name="fileName">The name of the file to get the size of.</param>
+        /// <returns>The size of the file in bytes.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when fileName is null.</exception>
+        /// <exception cref="ArgumentException">Thrown when fileName is empty.</exception>
+        /// <exception cref="FileNotFoundException">Thrown when the file does not exist.</exception>
         public static long GetFileSize(string fileName)
         {
+            if (fileName == null)
+                throw new ArgumentNullException(nameof(fileName));
+
+            if (string.IsNullOrEmpty(fileName))
+                throw new ArgumentException("File name cannot be null or empty.", nameof(fileName));
+
             var fileInfo = new FileInfo(fileName);
             return fileInfo.Length;
         }
